Fix issue where app would crash on first-run; make minor changes
When the user presses the record button and calls startRecording for the first time, startService(gpxService) does not wait for PermissionHandler.askWrite to return before starting the service and thus trying to write the gpxFile. This results in the app crashing as it has not yet obtained write permission, raising an uncaught error when trying to write a file.

To remedy this, startService(gpxServce) is placed in an else block so that the gpxService will only be started if write permission is granted. The downside is that the user must then press the record button a second time after granting write permissions in order to fully start recording.

Some minor changes were made for code styling consistency.